requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dgl-cu102==0.5.1
future==0.18.2
graphviz==0.14.1
idna==2.10
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
pandas==1.1.2
Pillow==7.2.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
torch==1.6.0
torchprof==1.1.1
torchsummary==1.5.1
torchvision==0.7.0
urllib3==1.25.10
